+++ PlexCleaner/SidecarFile.cs
@@ -44,7 +44,6 @@ public class SidecarFile
     private string _mkvMergeJson = string.Empty;
 
     private SidecarFileJsonSchema? _sidecarJson;
-    private StatesType _state;
 
     public SidecarFile(FileInfo mediaFileInfo)
     {
@@ -72,19 +71,7 @@ public class SidecarFile
     public MediaProps MkvMergeProps { get; private set; }
     public MediaProps MediaInfoProps { get; private set; }
 
-    public StatesType State
-    {
-        get => _state;
-        set
-        {
-            // Elevate loglevel
-            if (value != StatesType.None && value != _state)
-            {
-                PerFileLogLevel.Elevate();
-            }
-            _state = value;
-        }
-    }
+    public StatesType State { get; set; }
 
     public bool Create()
     {
@@ -321,8 +308,7 @@ public class SidecarFile
         MkvMergeProps = mkvMergeProps;
         FfProbeProps = ffProbeProps;
 
-        // Set state directly to prevent log elevation
-        _state = _sidecarJson.State;
+        State = _sidecarJson.State;
 
         return true;
     }